ShipSimulator2/requirements.txt
paho-mqtt==1.6.1
searoute==1.2.3